src/event_summarizer.py
import logging
from typing import Dict, List, Any
from llama_cpp import Llama

logger = logging.getLogger(__name__)

class EventSummarizer:
    def __init__(self, model_path: str = "models/llama-2-7b.gguf"):
        """
        Initialize the event summarizer with a local LLM model.
        
        Args:
            model_path (str): Path to the local LLM model file
        """
        try:
            self.llm = Llama(model_path=model_path)
        except Exception as e:
            logger.error(f"Failed to load LLM model: {str(e)}")
            raise
            
    def summarize_events(self, events_data: Dict[str, Any]) -> List[Dict[str, Any]]:
        """
        Summarize events using the LLM.
        
        Args:
            events_data (Dict): Raw event data from GDELT
            
        Returns:
            List of summarized events with locations and metadata
        """
        summarized_events = []
        
        for event in events_data.get("events", []):
            try:
                summary = self._generate_summary(event)
                if summary:
                    summarized_events.append(summary)
            except Exception as e:
                logger.error(f"Failed to summarize event: {str(e)}")
                continue
                
        return summarized_events
        
    def _generate_summary(self, event: Dict[str, Any]) -> Dict[str, Any]:
        """
        Generate a summary for a single event using the LLM.
        
        Args:
            event (Dict): Single event data
            
        Returns:
            Dict containing the summarized event with metadata
        """
        # TODO: Implement actual LLM summarization
        # This is a placeholder
        return {
            "summary": "Event summary placeholder",
            "location": {
                "lat": 0.0,
                "lon": 0.0
            },
            "timestamp": event.get("date", ""),
            "tone": event.get("tone", 0.0)
        } 